mcp-101-detect-colors/pyproject.toml
[project]
name = "mcp-101-detect-colors"
version = "0.1.0"
description = "MCP tool for analyzing dominant colors in images"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "mcp[cli]>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"